+++ index.js
@@ -6,3 +6,5 @@ module.exports.Responsive =
 module.exports.Responsive.utils = require("./build/responsiveUtils");
 module.exports.WidthProvider =
   require("./build/components/WidthProvider").default;
+module.exports.CrossGridProvider =
+  require("./build/CrossGridProvider").default;